change for loop to filter
