+++ public/style.css
@@ -11,7 +11,7 @@
   --background-color: rgb(60, 64, 67);
 }
 /* Main Things */
-* {
+*{
   margin: 0;
   padding: 0;
   box-sizing: border-box;
